refactor: move directory content retrieval endpoint to a new route file
This commit moves the `listFolderContent` route function for the
GET `/:siteName/folders/:path` endpoint into a separate `directory.js`
route file, renaming it to `listDirectoryContent` with a new endpoint,
GET `/:siteName/files/:path`.

The rationale for this refactor is that conceptually, this endpoint
is to be used for not just folders, but also for images, files, and
any other directory in the repo. Therefore, it does not make sense to
group it as a `folder` endpoint, but as a more general directory
endpoint.
