+++ middleware/auth.js
@@ -47,8 +47,10 @@ auth.get('/v1', noVerify)
 auth.get('/v1/sites/:siteName/homepage', verifyJwt)
 auth.post('/v1/sites/:siteName/homepage', verifyJwt)
 
+// Directory
+auth.get('/v1/sites/:siteName/files/:path', verifyJwt)
+
 // Folder pages
-auth.get('/v1/sites/:siteName/folders/:path', verifyJwt)
 auth.get('/v1/sites/:siteName/folders/all', verifyJwt)
 
 // Collection pages
