+++ routes/folders.js
@@ -7,23 +7,9 @@ const { attachReadRouteHandlerWrapper } = require('../middleware/routeHandler')
 
 // Import classes
 const { CollectionConfig } = require('../classes/Config')
-const { Directory, RootType, FolderType } = require('../classes/Directory');
+const { Directory, RootType } = require('../classes/Directory.js');
 
 const ISOMER_TEMPLATE_DIRS = ['_data', '_includes', '_site', '_layouts']
-
-// List pages and directories in folder
-async function listFolderContent (req, res, next) {
-    const { accessToken } = req
-    const { siteName, path } = req.params
-
-    const decodedPath = decodeURIComponent(path)
-
-    const IsomerDirectory = new Directory(accessToken, siteName)
-    const folderType = new FolderType(decodedPath)
-    IsomerDirectory.setDirType(folderType)
-    const directoryContents = await IsomerDirectory.list()
-    res.status(200).json({ directoryContents })
-}
 
 // List pages and directories from all folders
 async function listAllFolderContent (req, res, next) {
@@ -54,7 +40,6 @@ async function listAllFolderContent (req, res, next) {
     res.status(200).json({ allFolderContent })
 }
 
-router.get('/:siteName/folders/:path', attachReadRouteHandlerWrapper(listFolderContent))
 router.get('/:siteName/folders/all', attachReadRouteHandlerWrapper(listAllFolderContent))
 
 module.exports = router;
