+++ server.js
@@ -20,6 +20,7 @@ const sitesRouter = require('./routes/sites')
 const pagesRouter = require('./routes/pages')
 const collectionsRouter = require('./routes/collections')
 const collectionPagesRouter = require('./routes/collectionPages')
+const directoryRouter = require('./routes/directory')
 const foldersRouter = require('./routes/folders')
 const resourceRoomRouter = require('./routes/resourceRoom')
 const resourcesRouter = require('./routes/resources')
@@ -58,6 +59,7 @@ app.use('/v1/sites', sitesRouter)
 app.use('/v1/sites', pagesRouter)
 app.use('/v1/sites', collectionsRouter)
 app.use('/v1/sites', collectionPagesRouter)
+app.use('/v1/sites', directoryRouter)
 app.use('/v1/sites', foldersRouter)
 app.use('/v1/sites', resourceRoomRouter)
 app.use('/v1/sites', resourcesRouter)
